refactor: rename module to services
