+++ lib/storage_tables/engine.rb
@@ -3,7 +3,7 @@
 require "rails/engine"
 
 require "storage_tables"
-require "storage_tables/service/registry"
+require "storage_tables/services/registry"
 require "storage_tables/reflection"
 
 module StorageTables
